Change default npc uuid version
