+++ Towny/src/main/java/com/palmergames/bukkit/config/ConfigNodes.java
@@ -2981,7 +2981,7 @@ public enum ConfigNodes {
 	
 	ECO_ADVANCED_NPC_UUID_VERSION(
 			"economy.advanced.npc_uuid_version",
-			"2",
+			"4",
 			"",
 			"# The UUID version to use for non-player accounts. This is used so that economy plugins can more easily differentiate between player and NPC accounts.",
 			"# Set to -1 to disable modifying npc uuids."),
